clean out more files

+++ Processed Data/process_data.py
@@ -48,7 +48,7 @@ def create_separate():
     other_df.to_csv(os.path.join(output_dir, "other_retested.csv"), index=False)
 
 
-dict_to_walk = os.path.join(dir_path, 'Processed Data', 'Cleaned JSON', 'Failed', "WCAG A")
+dict_to_walk = os.path.join(dir_path, 'Processed Data', 'Cleaned JSON')
 
 def valid_signups():
 
@@ -66,18 +66,20 @@ def valid_signups():
                     with open(file_path, 'r', encoding='utf-8') as jsonfile:
                         audit_data = json.load(jsonfile)
                         if 'url' in audit_data and audit_data['url'] in to_array:
-                            """ output_file_path = os.path.join("Filtered Sign Up Audits", file)
-                            shutil.copy2(file_path, output_file_path)
-                            print(f"Copied: {file}") """
-                            check_diff.append(audit_data['url'])
+                            output_file_path = os.path.join("Filtered Sign Up Audits", file)
+                            if not os.path.isfile(output_file_path):
+                                shutil.copy2(file_path, output_file_path)
+                                print(f"Copied: {file}")
+                                check_diff.append(audit_data['url'])
 
                 except FileNotFoundError:
                     print(f"Warning: File not found {file}")
     
-    diff = np.setdiff1d(to_array, check_diff)
-    data_d = pd.DataFrame(diff, columns=['Values'])
-    data_d.to_csv("difference", index=False)
+    #diff = np.setdiff1d(to_array, check_diff)
+    #data_d = pd.DataFrame(diff, columns=['Values'])
+    #data_d.to_csv("difference", index=False)
 
+valid_signups()
 
 def edit_sting(url):
     url_name = url
@@ -116,7 +118,7 @@ def get_diff_audits(directory):
 
 difference_dir = os.path.join(dir_path, 'Processed Data', 'difference.csv')
 difference_dir = pd.read_csv(difference_dir, usecols = ['Values'])
-get_diff_audits(difference_dir)
+#get_diff_audits(difference_dir)
 
 #to_array = df['Website'].to_numpy()
 #print(len(to_array))
